LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/xhr/send-network-error-async-events.sub-expected.txt

ALERT: test started
ALERT: test stepped
ALERT: ready state is 4
ALERT: load ended

PASS XMLHttpRequest: The send() method: Fire a progress event named error when Network error happens (synchronous flag is unset) 
